Lambada.Eval

Description

A module for evaluating expressions in Lambada

Documentation

evaluate :: Expr -> Either String Expr

A function that evaluates an expression

eval :: Ctx -> Expr -> Either String Expr

A function that evaluates an expression in a given context

type Ctx = Map String Expr

A data type for storing variables and their values