Lambada.Lambada

Description

A higher level interface for Lambada

Documentation

evalLambada :: String -> Either String Expr

A function that parses and evaluates a given string

evalLambada' :: Ctx -> String -> Either String Expr

A function that parses and evaluates a given string in a context

evaluate :: Expr -> Either String Expr

A function that evaluates an expression

eval :: Ctx -> Expr -> Either String Expr

A function that evaluates an expression in a given context

lambada :: Parser Expr

Parse a Lambada string

parse :: Parser a -> String -> Result String a

Parse with the parser and return the result.

type Ctx = Map String Expr

A data type for storing variables and their values