Lambada.Parser

Description

The parser module for Lambada

Documentation

lambada :: Parser Expr

Parse a Lambada string

data LangInfo

A collection of information about a language

Constructors

LangInfo

Fields

keywords :: [String]

operators :: [String]

lambadaInfo :: LangInfo

Language information for Lambada

keyword :: String -> Parser String

Parse a word with lookahead

operator :: Parser String

Parse one of the operators

expr :: Parser Expr

Parse an expression

nonApp :: Parser Expr

Parse anything but application

letExpr :: Parser Expr

Parse a let-expression

var :: Parser Expr

Parse a variable

abstraction :: Parser Expr

Parse a lambda abstraction

application :: Parser Expr

Parse an application

data Result b a

The result data type, inspired by either, but made into its own type to redefine the show instance.

Constructors

Err b

Ok (a, b)

module Lambada.Expr