Parser.Char

Description

A module containing the most basic string parsers. These mostly parse a single character.

Documentation

type Parser a = ParseT String a

The string parser type.

item :: Parser Char

Get one item from the string input, and update the state accordingly.

eof :: Parser ()

Succeed if this is the end of input.

sat :: (Char -> Bool) -> Parser Char

Parse a character satisfying the predicate.

letter :: Parser Char

Parse a character satisfying the predicate.

digit :: Parser Char

Parse a character satisfying the predicate.

alphaNum :: Parser Char

Parse a character satisfying the predicate.

char :: Char -> Parser Char

Match the given character.

space :: Parser ()

Skip whitespace, note that this skips newlines as well.

spaces :: Parser ()

Skip whitespace, note that this skips newlines as well.

module Parser.Prim