Parser.Parse

Description

A module containing functions that let you parse strings or files with the parsers you create.

You can also use the test parsers to print the results directly, giving you an easier time seeing the results.

Documentation

parse :: Parser a -> String -> Result String a

Parse with the parser and return the result.

parseTest :: Show a => Parser a -> String -> IO ()

Parse a given string, and print the result.

parseFile :: Parser a -> FilePath -> IO (Result String a)

Parse a file with a given parser.

parseFileTest :: Show a => Parser a -> FilePath -> IO ()

Parse a given file, and print the result.

module Parser.String