Parser.ParseT

Description

The most general parser type, along with some low level functions used to manipulate the State of the parser.

Documentation

newtype ParseT b a

The parser type, which is imply a function from one state to either a state, or a state along with a result.

Constructors

P

Fields

runParser :: State b -> Result b a

modifyState :: (State b -> State b) -> ParseT b ()

A parser that applies the given function to its state.

getState :: ParseT b (State b)

A parser that returns its state.

setState :: State b -> ParseT b ()

A parser that replaces its state with a given one.

module Parser.State

module Parser.Result

module Control.Applicative

module Control.Monad