Parser.Prim

Description

A module defining a few primitive, general parsers. These can be used for any kind of input.

Documentation

label :: String -> ParseT b a -> ParseT b a

Label a parser for better error messages

(<?>) :: ParseT b a -> String -> ParseT b a infixl 0

The same as label, but with the arguments flipped

lookAhead :: ParseT b a -> ParseT b a

Get the result of parsing with the input, but without changing the state. No input is consumed.

try :: ParseT b a -> ParseT b a

Parse with the input, but pretend that no input has been consumed if it fails. This can be used when arbitray lookahead is needed.

choice :: [ParseT b a] -> ParseT b a

Parse with the given parsers, and return the result of the first one to succeed. Implemented with <|>.

unexpected :: ParseT b a -> ParseT b ()

Fail if the given parser succeeds

notFollowedBy :: ParseT i a -> ParseT i b -> ParseT i a

Fail if the first input fails, or the second succeeds.

manyTill :: ParseT i a -> ParseT i b -> ParseT i [a]

Parse with the first parser until the second succeeds. The result of the second is thrown away. You can use lookAhead if this is undesireable.

between :: ParseT i a -> ParseT i b -> ParseT i c -> ParseT i c

Parse with the first input, then the third, then the second. Can be used to implement parens and similar parsers.

module Parser.ParseT