Parser.Result

Description

A module for the Result data type

It is very similar to the Either data type, but with some new functions.

Documentation

data Result b a

The result data type, inspired by either, but made into its own type to redefine the show instance.

Constructors

Err (State b)

Ok (a, State b)

err :: Result b a -> Bool

Check that the result is as expected.

ok :: Result b a -> Bool

Check that the result is as expected.