Parser.State

Description

A module containing the State data type, which is used in Parser.ParseT to represent the changing state while parsing.

Documentation

data State a

The state data type It keeps track of the remaining input, the location, whether or not input has been consumed, and the error stack.

Constructors

State

Fields

inp :: a

loc :: Loc

consumed :: Bool

parseError :: ParseError a

data ParseError a

A data type used to represent an error while parsing. This is used to generate a trace of what a parser expected, what the parser got, and where the error occured.

Constructors

ParseError

Fields

actual :: a

expected :: [Label]

data Loc

The location data type. This is used to keep track of the location in the State data type.

type Label = String

A label used to explain what was expected when the the ParseError occured.

labelState :: Label -> State a -> State a

Add an error to the State stack with the given label.

defaultState :: FilePath -> a -> State a

Generates a state with the given filepath and input.

incCol :: State a -> State a

Function that edits the location of the State

incLine :: State a -> State a

Function that edits the location of the State

resetCol :: State a -> State a

Function that edits the location of the State

clearExpected :: State a -> State a

Clear the expected list

updateError :: State a -> State a

Update the remaining input in the error