Parser.String

Description

A module containing parsers that parse strings.

Documentation

string :: String -> Parser String

Match a given string.

lexeme :: Parser a -> Parser a

Parse with a parser, then ignore trailing whitespace.

intLit :: Parser Integer

Parse an integer literal that is not followed by a letter.

strLit :: Parser String

Parse a string literal.

identifier :: Parser String

Parse an identifier.

symbol :: String -> Parser String

Parse a given string, and skip trailing whitespace.

parens :: Parser a -> Parser a

Parse with the given parser, but with surrounding parentheses.

brackets :: Parser a -> Parser a

Parse with the given parser, but with surrounding brackets.

braces :: Parser a -> Parser a

Parse with the given parser, but with surrounding braces.

sepBy :: Parser a -> Parser b -> Parser [a]

Parse with the first parser multiple times, separated by the second parser.

parse (identifier `sepBy` (char '|')) "foo|bar|baz"

Or, if you would like to ignore whitespace after the separator

parse (identifier `sepBy` (symbol "|") "foo | bar | baz")

commaSep :: Parser a -> Parser [a]

Parse multiple times with the given parser, separated by commas.

parse (commaSep intLit) "13,1,5,12"

list :: Parser a -> Parser [a]

Parse multiple times with a given parser, separated by commas, and surrounded by brackets.

parse (list intLit) "[123,5,1]"

tuple :: Parser a -> Parser [a]

Parse multiple times with a given parser, separated by commas, and surrounded by parentheses.

parse (tuple identifier) "(foo, bar, baz)"

mapTo :: Parser a -> Parser b -> Parser [(a, b)]

With surrounding braces, parse multiple occurences of the first parser, then a colon, then the second parser. These occurences are separated by commas.

parse (identifier `mapTo` var) "{abc : 123, xyz : 512}"

module Parser.Char