Types.Env

Description

An environment type that represents a map of variables and their values.

Documentation

data Env

The environment type

emptyEnv :: Env

An empty environment

insertEnv :: String -> Expr -> Env -> Env

Insert a key and a value into the environment

lookupEnv :: String -> Env -> Either String Expr

Look up the value of a key in the environment